0.04
Dodano formatowanie walut
Poprawiono formuły przeliczania ulepszeń i przychodów (poprawić wartości
domyślne)

+++ Farmer/Assets/Scripts/Classes/UpgradeButton.cs
@@ -20,7 +20,7 @@ public class UpgradeButton : MonoBehaviour {
     /// Metoda zwraca koszt budynku
     /// </summary>
     /// <returns></returns>
-    public int GetBuildingCost()
+    public BigInteger GetBuildingCost()
     {
         return Helper.GetBuildingManager().GetBuildingByType(Type).GetCost();
     }
